fix: #8 #9 #14 - changes to deletion:
1) don't take backup when deleting server
2) don't send any deletion success messages - channel is (about to be) deleted
3) get name name from channel context rather than requiring it to be provided

+++ bot/cogs/admin.py
@@ -5,6 +5,7 @@ from discord.ext import commands
 from ..services import channel_mapping_service, game_service, ip_service
 from ..utilities import random_string
 from .roles import FACTORIO_CATEGORY
+from ..helpers import game_mapping_helper
 
 
 name_pattern = re.compile("^[A-Za-z][A-Za-z0-9-]*$")
@@ -38,38 +39,37 @@ class Admin(commands.Cog):
                        " :partying_face:")
 
     @commands.command(help='Delete a game',
-                      description="This will permanently delete the game (a backup will be taken" +
-                      "first).")
-    async def delete(self, ctx, name, confirmation_phrase=None):
-        if (confirmation_phrase is not None and
-                confirmation_phrase == self.confirmation_phrases.get(name)):
-            del self.confirmation_phrases[name]
-            await ctx.send(f'Deleting {name}')
-            await game_service.delete_game(name)
-            await ctx.send(
-                f'Your complimentary backup can be downloaded using `!list-backups {name}`')
-            # We want to remove any channel associations with this game - the easiest way to do that
-            # is just to validate the mappings again
-            await channel_mapping_service.validate_mappings(self.bot)
-            for guild in self.bot.guilds:
-                category = discord.utils.get(
-                    guild.categories, name=FACTORIO_CATEGORY)
-                channel = discord.utils.get(category.channels, name=name)
-                if channel is not None:
-                    await channel.delete()
-            await ctx.send(f'Deleted {name}! :skull_crossbones:')
-        elif confirmation_phrase is None or self.confirmation_phrases.get(name) is None:
-            self.confirmation_phrases[name] = random_string(10)
-            await ctx.send(':warning: :warning: :warning: The game and associated discord channel' +
-                           " will be permanently deleted. If you want to host it again you'll " +
-                           " have to set that up manually - I cannot help you with that (although" +
-                           " I will provide you with the backup). :warning: :warning: :warning:" +
-                           "\nTo confirm the delete, use " +
-                           f"`!delete {name} {self.confirmation_phrases[name]}`")
-        else:
-            self.confirmation_phrases[name] = random_string(10)
-            await ctx.send(':no_entry_sign: Confirmation phrase did not match - to confirm the ' +
-                           f'delete, use `!delete {name} {self.confirmation_phrases[name]}`')
+                      description="Permanently delete the game.")
+    async def delete(self, ctx, confirmation_phrase=None):
+        name = await game_mapping_helper.game_from_context(ctx, self.bot)
+        if name is not None:
+            if (confirmation_phrase is not None and
+                    confirmation_phrase == self.confirmation_phrases.get(name)):
+                del self.confirmation_phrases[name]
+                await ctx.send(f'Deleting {name}')
+                await game_service.delete_game(name)
+                # We want to remove any channel associations with this game - the
+                # easiest way to do that is just to validate the mappings again
+                await channel_mapping_service.validate_mappings(self.bot)
+                for guild in self.bot.guilds:
+                    category = discord.utils.get(
+                        guild.categories, name=FACTORIO_CATEGORY)
+                    channel = discord.utils.get(category.channels, name=name)
+                    if channel is not None:
+                        await channel.delete()
+            elif confirmation_phrase is None or self.confirmation_phrases.get(name) is None:
+                self.confirmation_phrases[name] = random_string(10)
+                await ctx.send(f':warning: :warning: :warning: Game "{name}" and associated ' +
+                               'discord channel will be permanently deleted. If you want to ' +
+                               "host it again you'll have to set that up manually. **Make sure " +
+                               "you have taken the appropriate backups.** :warning: :warning: " +
+                               ":warning: \nTo confirm the delete, use " +
+                               f"`!delete {self.confirmation_phrases[name]}`")
+            else:
+                self.confirmation_phrases[name] = random_string(10)
+                await ctx.send(':no_entry_sign: Confirmation phrase did not match - to confirm ' +
+                               'the delete, use ' +
+                               f'`!delete {self.confirmation_phrases[name]}`')
 
     @commands.command(name='set-game',
                       help='Link the current channel to the specified game',
